Implemented restrictions based on roles and permissions

+++ app/Http/Controllers/CommentController.php
@@ -25,6 +25,11 @@ class CommentController extends Controller
 
     public function destroy(Comment $comment)
     {
+        if($comment->user_id !== Auth::id())
+        {
+            abort(403);
+        }
+
         $featureID = $comment->feature_id;
         $comment->delete();
